Move onchain-restore methods to own module
