+++ lib/core/src/lib.rs
@@ -170,6 +170,7 @@ pub mod logger;
 pub mod model;
 pub mod persist;
 pub(crate) mod receive_swap;
+mod restore_onchain;
 pub mod sdk;
 pub(crate) mod send_swap;
 pub(crate) mod swapper;
